Custom domains and separation of plugin settings

+++ amazon-link-engine/trunk/amazon-link-engine.php
@@ -3,10 +3,14 @@
 Plugin Name: Amazon Link Engine
 Plugin URI:
 Description: Automatically optimizes Amazon product links for your global audience and allows you to earn commissions on sales.
-Version: 1.1.9
+Version: 1.2.2
 Author: GeoRiot Networks, Inc.
 Author URI: http://geni.us
 */
+
+//Change this if you need to run a migration (eg change setting names, dbm etc). See genius_ale_update_db_check()
+global $genius_ale_db_version;
+$genius_ale_db_version = '1.1';
 
 if (!defined('WP_CONTENT_URL'))
       define('WP_CONTENT_URL', get_option('siteurl').'/wp-content');
@@ -19,46 +23,76 @@ if (!defined('WP_PLUGIN_DIR'))
 
 
 // OPTIONS
+// Will set to old values if the new genius ones don't already exist in the db, and old georiot ones are available
+function activate_genius_autolinker() {
+  global $genius_ale_db_version;
 
-function activate_georiot_autolinker() {
-  add_option('georiot_tsid', '');
-  add_option('georiot_api_key', '');
-  add_option('georiot_api_secret', '');
-  add_option('georiot_api_remind', 'yes');
-  add_option('georiot_preserve_tracking', 'no');
+  add_option('genius_ale_domain', '');
+  add_option('genius_ale_tsid', '');
+  add_option('genius_ale_api_key', '');
+  add_option('genius_ale_api_secret', '');
+  add_option('genius_ale_api_remind', 'yes');
+  add_option('genius_ale_preserve_tracking', 'no');
+  add_option('genius_ale_db_version', $genius_ale_db_version);
 }
 
-function deactivate_georiot_autolinker() {
-  delete_option('georiot_tsid');
-  delete_option('georiot_api_key');
-  delete_option('georiot_api_secret');
-  delete_option('georiot_api_remind');
-  delete_option('georiot_preserve_tracking');
+function deactivate_genius_autolinker() {
+  delete_option('genius_ale_domain');
+  delete_option('genius_ale_tsid');
+  delete_option('genius_ale_api_key');
+  delete_option('genius_ale_api_secret');
+  delete_option('genius_ale_api_remind');
+  delete_option('genius_ale_preserve_tracking');
+  delete_option('genius_ale_db_version');
 }
 
-function admin_init_georiot_autolinker() {
-  register_setting('amazon-link-engine', 'georiot_tsid');
-  register_setting('amazon-link-engine', 'georiot_api_key');
-  register_setting('amazon-link-engine', 'georiot_api_secret');
-  register_setting('amazon-link-engine', 'georiot_api_remind');
-  register_setting('amazon-link-engine', 'georiot_preserve_tracking');
+function admin_init_genius_autolinker() {
+  register_setting('amazon-link-engine', 'genius_ale_domain');
+  register_setting('amazon-link-engine', 'genius_ale_tsid');
+  register_setting('amazon-link-engine', 'genius_ale_api_key');
+  register_setting('amazon-link-engine', 'genius_ale_api_secret');
+  register_setting('amazon-link-engine', 'genius_ale_api_remind');
+  register_setting('amazon-link-engine', 'genius_ale_preserve_tracking');
+  register_setting('amazon-link-engine', 'genius_ale_db_version');
 }
 
 
-function admin_menu_georiot_autolinker() {
-  add_options_page('Amazon Link Engine', 'Amazon Link Engine', 'manage_options', 'amazon-link-engine', 'options_page_georiot_autolinker');
+//Backwards compatibility: Migrate old vals to new ones
+function genius_ale_migrate_1() {
+  global $genius_ale_db_version;
+
+  update_option('genius_ale_tsid', get_option('georiot_tsid'));
+  update_option('genius_ale_api_key', get_option('georiot_api_key'));
+  update_option('genius_ale_api_secret', get_option('georiot_api_secret'));
+  update_option('genius_ale_api_remind', get_option('georiot_api_remind'));
+  update_option('genius_ale_preserve_tracking', get_option('georiot_preserve_tracking'));
+  update_option('genius_ale_db_version', $genius_ale_db_version);
+
+  //Delete the obsolete values, only if the old iTunes plugin isn't installed
+  if( !function_exists( 'georiot_ile' ) ) {
+    delete_option('georiot_tsid');
+    delete_option('georiot_api_key');
+    delete_option('georiot_api_secret');
+    delete_option('georiot_api_remind');
+    delete_option('georiot_preserve_tracking');
+  }
 }
 
-function options_page_georiot_autolinker() {
+
+function admin_menu_genius_autolinker() {
+  add_options_page('Amazon Link Engine', 'Amazon Link Engine', 'manage_options', 'amazon-link-engine', 'options_page_genius_autolinker');
+}
+
+function options_page_genius_autolinker() {
   include(WP_PLUGIN_DIR.'/amazon-link-engine/options.php');
 }
 
 
 
 // Show notice in dashboard home page and plugin page if API isn't connected
-function georiot_admin_notice(){
+function genius_admin_notice(){
   if (strpos($_SERVER['PHP_SELF'],'wp-admin/index.php') !== false  || strpos($_SERVER['PHP_SELF'],'wp-admin/plugins.php') !== false ) {
-    if (get_option('georiot_api_remind') == 'yes' && get_option('georiot_tsid') == '') {
+    if (get_option('genius_ale_api_remind') == 'yes' && get_option('genius_ale_tsid') == '') {
       ?>
       <div class="update-nag">
         <p><?php _e('<strong>Your Amazon Link Engine plugin is installed and working.</strong> <br>To use reporting and commissions, <a href="'.admin_url().'options-general.php?page=amazon-link-engine">enter your GeniusLink API values.</a>. Or, you can <a href="'.admin_url().'options-general.php?page=amazon-link-engine">disable this reminder.</a>'); ?></p>
@@ -68,19 +102,25 @@ function georiot_admin_notice(){
   }
 }
 
-// BEGIN FUNCTION TO SHOW GEORIOT JS
+// BEGIN FUNCTION TO SHOW GENIUS JS
 
-function georiot_autolinker() {
+function genius_ale() {
 
-  if (get_option('georiot_tsid') == '') {
+  if (get_option('genius_ale_tsid') == '') {
     $gr_use_tsid = 4632;
   } else {
-    $gr_use_tsid = get_option('georiot_tsid');
+    $gr_use_tsid = get_option('genius_ale_tsid');
+  }
+
+  if (get_option('genius_ale_domain') != 'geni.us' && get_option('genius_ale_domain') != '') {
+    $gr_use_domain = ", 'http://" . get_option("genius_ale_domain")."'" ;
+  } else {
+    $gr_use_domain = '';
   }
 
   $preserve_tracking = 'false';
 
-  if (get_option('georiot_preserve_tracking') == 'yes') {
+  if (get_option('genius_ale_preserve_tracking') == 'yes') {
     $preserve_tracking = 'true';
   }
 
@@ -89,35 +129,52 @@ function georiot_autolinker() {
   <script src="//cdn.georiot.com/snippet.js"></script>
   <script type="text/javascript">
     jQuery(document).ready(function( $ ) {
-      var tsid = <?php echo $gr_use_tsid ?>;
-      Georiot.amazon.convertToGeoRiotLinks(tsid, <?php print($preserve_tracking)?>);
+      Georiot.amazon.convertToGeoRiotLinks(<?php echo $gr_use_tsid ?>, <?php print($preserve_tracking)?><?php print($gr_use_domain) ?>);
     });
   </script>
 <?php
 }
-// END FUNCTION TO SHOW GEORIOT JS
+// END FUNCTION TO SHOW GENIUS JS
 
-register_activation_hook(__FILE__, 'activate_georiot_autolinker');
-register_deactivation_hook(__FILE__, 'deactivate_georiot_autolinker');
+register_activation_hook(__FILE__, 'activate_genius_autolinker');
+register_deactivation_hook(__FILE__, 'deactivate_genius_autolinker');
 
 if (is_admin()) {
-  add_action('admin_init', 'admin_init_georiot_autolinker');
-  add_action('admin_menu', 'admin_menu_georiot_autolinker');
-  add_action('admin_notices', 'georiot_admin_notice');
+  add_action('admin_init', 'admin_init_genius_autolinker');
+  add_action('admin_menu', 'admin_menu_genius_autolinker');
+  add_action('admin_notices', 'genius_admin_notice');
 }
 
 if (!is_admin()) {
-  add_action('wp_head', 'georiot_autolinker');
+  add_action('wp_head', 'genius_ale');
 }
+
+
+//Update the plugin if needed
+function genius_ale_update_db_check() {
+  global $genius_ale_db_version;
+  $current_ale_db_version = get_option('genius_ale_db_version');
+
+  if ( $current_ale_db_version != $genius_ale_db_version ) {
+
+    //Check if they are on the oldest version of the genius plugin db
+    if( !$current_ale_db_version ) {
+      genius_ale_migrate_1();
+    }
+  }
+}
+
+
+add_action( 'plugins_loaded', 'genius_ale_update_db_check' );
 
 
 // SHOW SETTINGS OPTION IN THE PLUGIN PAGE
 // Settings link
-function georiot_add_settings_link($actions) {
-  $actions = array('settings' => sprintf('<a href="%s" title="%s">%s</a>', admin_url().'options-general.php?page=amazon-link-engine', __('Configure GeoRiot Plugin.'), __('Settings'))) + $actions;
+function genius_add_settings_link($actions) {
+  $actions = array('settings' => sprintf('<a href="%s" title="%s">%s</a>', admin_url().'options-general.php?page=amazon-link-engine', __('Configure this plugin'), __('Settings'))) + $actions;
   return $actions;
 }
-add_filter('plugin_action_links_' . plugin_basename(__FILE__), 'georiot_add_settings_link');
+add_filter('plugin_action_links_' . plugin_basename(__FILE__), 'genius_add_settings_link');
 
 
 ?>
